Add navigation between article list screen to about screen, using voyager
