+++ composeApp/src/commonMain/kotlin/App.kt
@@ -6,21 +6,18 @@ import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import application.ArticlesScreen
-import dev.icerock.moko.mvvm.compose.getViewModel
-import dev.icerock.moko.mvvm.compose.viewModelFactory
+import ui.article.ArticlesScreen
+import cafe.adriel.voyager.navigator.Navigator
 
 @Composable
 fun App() {
     MaterialTheme {
-        val articleViewModel = getViewModel(Unit, viewModelFactory { ArticlesViewModel() })
-
         Surface(
             modifier = Modifier.fillMaxSize(),
             color = MaterialTheme.colorScheme.background
         ) {
             Column(Modifier.fillMaxWidth(), horizontalAlignment = Alignment.CenterHorizontally) {
-                ArticlesScreen(onAboutButtonClick = {},articleViewModel)
+                Navigator(ArticlesScreen())
             }
         }
     }
