+++ composeApp/src/commonMain/kotlin/ui/about/AboutScreenContent.kt
@@ -1,33 +1,51 @@
+package ui.about
+
+import Platform
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.Divider
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 
 
 @Composable
-fun AboutScreen(
+fun AboutScreenContent(
+    onUpButtonClick: () -> Unit
 ) {
     Column {
-        Toolbar()
+        Toolbar(onUpButtonClick)
         ContentView()
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3Api::class)
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 private fun Toolbar(
+    onUpButtonClick: () -> Unit
 ) {
     TopAppBar(
         title = { Text(text = "About Device") },
+        navigationIcon = {
+            IconButton(onClick = onUpButtonClick) {
+                Icon(
+                    imageVector = Icons.Default.ArrowBack,
+                    contentDescription = "Up Button",
+                )
+            }
+        }
     )
 }
 
@@ -64,7 +82,7 @@ private fun RowView(
         Text(
             text = title,
             style = MaterialTheme.typography.bodySmall,
-            color = androidx.compose.ui.graphics.Color.Gray,
+            color = Color.Gray,
         )
         Text(
             text = subtitle,
